# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Products"
                ],
                "summary": "List all products",
                "parameters": [
                    {
                        "description": "Pagination parameters",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_shared.Pagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product to create",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.CreateProductParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a product by its ID",
                "tags": [
                    "Admin Products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update product details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Products"
                ],
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated product details",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.UpdateProductParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a product by its ID",
                "tags": [
                    "Admin Products"
                ],
                "summary": "Delete a product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-wordpress_internal_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-wordpress_internal_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "go-wordpress_internal_shared.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
...
        "go-wordpress_internal_storage_sql_sqlc.CreateProductParams": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer",
                    "format": "int32"
                },
                "description": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "image": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "link": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "format": "int64"
                },
                "status": {
                    "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.EntityStatus"
                },
                "title": {
                    "type": "string"
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "go-wordpress_internal_storage_sql_sqlc.EntityStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "in_progress"
            ],
            "x-enum-varnames": [
                "EntityStatusActive",
                "EntityStatusInactive",
                "EntityStatusInProgress"
            ]
        },
...
        "go-wordpress_internal_storage_sql_sqlc.Product": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer",
                    "format": "int32"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "image": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "link": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "format": "int64"
                },
                "status": {
                    "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.EntityStatus"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "go-wordpress_internal_storage_sql_sqlc.UpdateProductParams": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer",
                    "format": "int32"
                },
                "description": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "image": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "link": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "format": "int64"
                },
                "status": {
                    "$ref": "#/definitions/go-wordpress_internal_storage_sql_sqlc.EntityStatus"
                },
                "title": {
                    "type": "string"
                },
                "websiteID": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}